+++ starkbank/dynamicbrcode/__init__.py
@@ -1,1 +1,2 @@
+from .rule.__rule import Rule
 from .__dynamicbrcode import create, get, query, page